Fix sign comparison

+++ platform/io/file.c
@@ -14,7 +14,7 @@ void file_text_create_from_cstr_if_different(const Path p, const char* str)
   const usize len = strlen(str);
 
   const ssize file_size = _file_size(p.cstr);
-  if(file_size>=0 && file_size == len)
+  if(file_size>=0 && (usize)file_size == len)
   {
     const Mem_Region prev_stack = STACK;
     Bytes_Mut bytes = _read_all_file_bytes(p.cstr, &STACK);
